:bug: Introduce timings, fix port mapping, separate master/slave
First of all we fixed port mappings that were wrong due to copy pasting.
Then we learnt that waiting some time for the electrical signal to stabilize
is actually a good idea.
Then we also understood that we needed more sepeartion in the API between
master and slave. This was because we were controlling SCL line in the slave
directly which is not what the protocol specifies to do. At this moment
the library is still not functional. The main problem now is getting the
timing right for acknowledging ACKs primarly, this is the issue that we are
facing right now.