:sparkles: It works (kinda), timings issues as per usual
I had it working but suddenly removing some calls to print for debugging
it all stopped working. My take on this is that those prints where introducing
the exact amount of delay that was necessary for the master/slave to be in sync.
I'm really not sure how to approach this synchronization issue right now.